requirements.txt
autopep8==1.4.4
backcall==0.1.0
decorator==4.4.0
Django==2.2.6
django-debug-toolbar==2.0
django-extensions==2.2.5
django-jalali==3.1.0
django-taggit==1.1.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
ipython==7.9.0
ipython-genutils==0.2.0
jdatetime==3.6.2
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.3
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
